fix: address all review comments
- Extract XML builder to src/util/camt053-builder.ts (#6)
- Add IBAN validation via Validations.Iban for both IBAN fields (#2)
- Escape bookingDate/valueDate/currency in XML output (#3)
- Add NaN guard for parseFloat on amount (#5)
- Use crypto.randomUUID() for AcctSvcrRef (#8)
- Change button label from "Next" to "Upload" (#10)
